% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CA_ClimateNormals}
\alias{CA_ClimateNormals}
\title{California Climate Normals}
\format{
A data frame with 5100 observations and 51 variables using mm for precipitation and Celsius for temperature variables
}
\source{
\url{https://www.ncdc.noaa.gov/}
}
\usage{
CA_ClimateNormals
}
\description{
30 year normals of monthly temperature, precipitation, and other climate variables
from NOAA/NCDC weather stations in California
}
\keyword{datasets}
