% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eucoakrainfallrunoffTDR}
\alias{eucoakrainfallrunoffTDR}
\title{Eucalyptus and Oak rainfall, runoff, soil moisture}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 90 rows and 17 columns.
}
\source{
measurements collected for
Thompson, Davis, Oliphant (2016) Surface runoff and soil erosion under
eucalyptus and oak canopy. Earth Surface Processes and Landforms 41: 1018-1026.
}
\usage{
eucoakrainfallrunoffTDR
}
\description{
Study of runoff under Eucalyptus and oak canopy in Bay Area sites
}
\keyword{datasets}
