% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{ex}
\alias{ex}
\title{Access external data in package}
\usage{
ex(dta)
}
\arguments{
\item{dta}{filename to access from extdata, including folders}
}
\value{
path to the file
}
\description{
Access external data in package
}
\examples{
read.csv(ex("sierra/sierraStations.csv"))
}
