% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sierraClimate}
\alias{sierraClimate}
\title{Sierra climate data}
\format{
A data frame with 983 observations and 53 variables
same as California Climate Normals, with county and geometry added
}
\usage{
sierraClimate
}
\description{
Derived from monthly California Climate Normals (in mm, degC) at Sierra Stations,
with County names added from a spatial join.
}
\keyword{datasets}
