% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sierraFeb}
\alias{sierraFeb}
\title{Sierra February climate data}
\format{
A data frame with 82 entries and 7 variables selected and renamed \describe{
  \item{STATION_NAME}{Station name}
  \item{COUNTY}{County Name}
  \item{ELEVATION}{Elevation in meters}
  \item{LATITUDE}{Latitude in decimal degrees}
  \item{LONGITUDE}{Longitude in decimal degrees}
  \item{PRECIPITATION}{February Average Precipitation in mm}
  \item{TEMPERATURE}{Febrary Average Temperature in degrees C}
}
}
\source{
\url{https://www.ncdc.noaa.gov/}
}
\usage{
sierraFeb
}
\description{
Selection from SierraData to only include February data
}
\keyword{datasets}
