% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sierraStations}
\alias{sierraStations}
\title{Sierra Stations}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 82 rows and 2 columns.
}
\source{
\url{https://www.ncdc.noaa.gov/}
}
\usage{
sierraStations
}
\description{
Station names and ids selected to represent Sierra Nevada, southern Sierra, and towns
adjacent the Sierra in the central valley and east side.  Selected from NCDC points.
}
\keyword{datasets}
