% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilCO2_97}
\alias{soilCO2_97}
\title{Soil CO2 1997}
\format{
data frame of 558 observations of 5 variables \describe{
   \item{SITE}{Site number north to south}
   \item{DATE}{Date}
   \item{TIME}{Time}
   \item{TEMPC}{Temperature degrees C}
   \item{CO2}{Soil CO2 concentration in percent}
 }
}
\source{
field measurements from 1997 research by Davis, Amato, and Kiefer
}
\usage{
soilCO2_97
}
\description{
Soils CO2 measurements from a transect of 11 sites from Super Sink across Marble Valley
in the Marble Mountains Wilderness of California
}
\keyword{datasets}
