% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRI_1997_CA}
\alias{TRI_1997_CA}
\title{EPA Toxic Release Inventory data, 1997, CA}
\format{
data frame with 4,198 entries, 110 variables (see the web page)
}
\source{
:\url{https://www.epa.gov/toxics-release-inventory-tri-program/tri-data-and-tools}
}
\usage{
TRI_1997_CA
}
\description{
EPA Toxic Release Inventory (TRI) for 1997, sites in California.
}
\keyword{datasets}
