% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRI_2017_CA}
\alias{TRI_2017_CA}
\title{EPA Toxic Release Inventory data, 2017, CA}
\format{
data frame with 3,685 entries, 110 variables (see the web page)
}
\source{
:\url{https://www.epa.gov/toxics-release-inventory-tri-program/tri-data-and-tools}
}
\usage{
TRI_2017_CA
}
\description{
EPA Toxic Release Inventory (TRI) for 2017, sites in California.
}
\keyword{datasets}
