% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XSptsNDVI}
\alias{XSptsNDVI}
\title{Meadow cross section with vegetation and NDVI}
\format{
data frame with 29 observations and 6 variables \describe{
  \item{DistNtoS}{cross sectional distance location in m}
  \item{elevation}{laser-leveled elevation of data point}
  \item{vegetation}{vegetation type}
  \item{geometry}{UTM zone 10 N coordinates of point}
  \item{NDVIgrowing}{NDVI from drone imagery during growing season Jul 2017}
  \item{NDVIsenescent}{NDVI from drone imagery during senescent season Sep 2018}
  }
}
\source{
from study of Knuthson meadow for
Davis et al. (2020). Unpiloted aerial system (UAS)-supported biogeomorphic
analysis of restored Sierra Nevada montane meadows. Remote Sensing 12: 1828. 25p.
}
\usage{
XSptsNDVI
}
\description{
NDVI derived from multispectral drone imagery of Knuthson Meadow cross section
during growing and senescent seasons in 2017 & 2018.
}
\keyword{datasets}
