% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_eucoak}
\alias{tidy_eucoak}
\title{Eucalyptus and Oak rainfall and runoff, tidied}
\format{

}
\source{
measurements collected for
Thompson, Davis, Oliphant (2016) Surface runoff and soil erosion under
eucalyptus and oak canopy. Earth Surface Processes and Landforms 41: 1018-1026.
}
\usage{
tidy_eucoak
}
\description{
Eucalyptus and Oak rainfall and runoff, tidied
}
\keyword{datasets}
